0610 Js Practice and Subject
Added:
0610/
0610/js_01.html
0610/js_02.html
0610/js_03.html
0610/js_04.html
0610/js_05.html
0610/js_02_add.html
0610/js_03_add.html
0610/js_04_add.html
0610/js_05_add.html
0610/light-bulb1.png
0610/light-bulb2.png
0610/new.html
0610/result.html

This Commit contained:
This commit contained js practice and subject. The practice was about js basic and js event.
The subject was about js calculator and light bulb switch.
The light bulb switch was about changing the image when the button is clicked.
The calculator was about making a calculator with js.
The Heading Example was about making a heading with js.
The Table Example was about making a table with js.

Signed-off-by: tmddn3070 <tmddn3070@gmail.com>